fix: resolve MISRA.ETYPE.ASSIGN.2012 violations
Change struct members to bool type where semantically correct.
Fix macro types and add explicit casts for narrowing conversions.
Update constant suffixes and memset usage to match codebase patterns.

All changes are type-safety improvements with no functional impact.

+++ devices/TPS6522x-Q1/src/pmic_io.c
@@ -188,7 +188,7 @@ int32_t Pmic_ioTxByte(const Pmic_Handle_t *handle, uint16_t regAddr, uint8_t txD
         return status;
     }
 
-    page = (regAddr >> SPI_ADDR_BYTE_SHIFT) & SPI_PAGE_MASK;
+    page = (uint8_t)((regAddr >> SPI_ADDR_BYTE_SHIFT) & SPI_PAGE_MASK);
 
     // SPI MODE
     if (handle->commMode == PMIC_INTF_SPI)
@@ -310,7 +310,7 @@ int32_t Pmic_ioRxByte(const Pmic_Handle_t *handle, uint16_t regAddr, uint8_t *rx
         return PMIC_ST_ERR_NULL_PARAM;
     }
 
-    page = (regAddr >> SPI_ADDR_BYTE_SHIFT) & SPI_PAGE_MASK;
+    page = (uint8_t)((regAddr >> SPI_ADDR_BYTE_SHIFT) & SPI_PAGE_MASK);
 
     // SPI MODE
     if (handle->commMode == PMIC_INTF_SPI)
